fix: moved opensea to api & fixed env var
