+++ .env.sample
@@ -17,6 +17,7 @@ NEXT_PUBLIC_VERCEL_URL=http://localhost:3000
 WEB3_STORAGE_TOKEN=eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJzdWIiOiJkaWQ6ZXRocjoweGE0OWIyMGIyMzVGY2E3N0QzRURlZWFDMzlDODkyZkVENmUzOTU5OTQiLCJpc3MiOiJ3ZWIzLXN0b3JhZ2UiLCJpYXQiOjE2MzA3MDMyMDg0ODEsIm5hbWUiOiJNeU1ldGEifQ.q7_i_XmkIMb6_6u9pNI6tkNUq0vhiJ0e1oV8GLdeez0
 NEXT_PUBLIC_CERAMIC_URL=https://ceramic-clay.3boxlabs.com
 NEXT_PUBLIC_CERAMIC_NETWORK=testnet-clay
+OPENSEA_API_KEY=    # add to view users' nfts in profiles
 
 # packages/discord-bot
 GRAPHQL_URL=http://localhost:8080/v1/graphql
